+++ optics/src/main/scala/com/calclavia/edx/optics/util/FortronUtility.scala
@@ -25,7 +25,7 @@ object FortronUtility {
 				transferMode match {
 					case TransferMode.equalize =>
 						frequencyBlocks.foreach(machine => {
-							val capacityPercentage = machine.getFortronCapacity.toDouble / totalCapacity.toDouble
+							val capacityPercentage = machine.getFortronCapacity / totalCapacity
 							val amountToSet = (totalFortron * capacityPercentage).toInt
 							doTransferFortron(source, machine, amountToSet - machine.getFortron, limit)
 						})
@@ -92,8 +92,8 @@ object FortronUtility {
 				toBeInjected = transferer.addFortron(receiver.removeFortron(toBeInjected, true), true)
 				if (Side.get().isClient && toBeInjected > 0 && !isCamo) {
 					val particle = world.addClientEntity(new EntityMagneticBeam(FieldColor.blue, 20))
-					particle.setPosition(block.position.toDouble + 0.5)
-					particle.setTarget(receiver.asInstanceOf[Block].position.toDouble + 0.5)
+					particle.setPosition(block.position + 0.5)
+					particle.setTarget(receiver.asInstanceOf[Block].position + 0.5)
 				}
 			}
 			else {
@@ -102,8 +102,8 @@ object FortronUtility {
 				toBeEjected = receiver.addFortron(transferer.removeFortron(toBeEjected, true), true)
 				if (Side.get().isClient && toBeEjected > 0 && !isCamo) {
 					val particle = world.addClientEntity(new EntityMagneticBeam(FieldColor.blue, 20))
-					particle.setTarget(block.position.toDouble + 0.5)
-					particle.setPosition(receiver.asInstanceOf[Block].position.toDouble + 0.5)
+					particle.setTarget(block.position + 0.5)
+					particle.setPosition(receiver.asInstanceOf[Block].position + 0.5)
 				}
 			}
 		}
